# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bible Verse Daily API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Verse {
  id: number;
  book: string;
  chapter: number;
  verseNumber: number;
  text: string;
  version: string;
}

export interface DailyVerse {
  verse: Verse;
  reflection: string;
  date: string;
}

export interface VerseList {
  verses: Verse[];
  total: number;
}

export interface Devotional {
  id: number;
  title: string;
  content: string;
  verseReference: string;
  verseText: string;
  category: string;
  readTime: number;
  date: string;
}

export interface DevotionalList {
  devotionals: Devotional[];
  total: number;
}

export interface Book {
  name: string;
  testament: string;
  chapters: number;
}

export interface BookList {
  books: Book[];
}

export type GetVersesParams = {
  book?: string;
  chapter?: number;
  limit?: number;
  offset?: number;
};

export type GetDevotionalsParams = {
  category?: string;
  limit?: number;
  offset?: number;
};