# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bible Verse Daily API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: verses
    description: Bible verse operations
  - name: devotionals
    description: Devotional operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /verses/daily:
    get:
      operationId: getDailyVerse
      tags: [verses]
      summary: Get the daily verse
      description: Returns today's verse of the day
      responses:
        "200":
          description: Daily verse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyVerse"
  /verses:
    get:
      operationId: getVerses
      tags: [verses]
      summary: Get verses
      description: Returns a list of verses, optionally filtered by book/chapter
      parameters:
        - name: book
          in: query
          schema:
            type: string
        - name: chapter
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of verses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerseList"
  /verses/{id}:
    get:
      operationId: getVerse
      tags: [verses]
      summary: Get a specific verse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: A verse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Verse"
  /devotionals:
    get:
      operationId: getDevotionals
      tags: [devotionals]
      summary: Get devotionals
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of devotionals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DevotionalList"
  /devotionals/{id}:
    get:
      operationId: getDevotional
      tags: [devotionals]
      summary: Get a specific devotional
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: A devotional
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Devotional"
  /books:
    get:
      operationId: getBooks
      tags: [verses]
      summary: Get list of Bible books
      responses:
        "200":
          description: List of books
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookList"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Verse:
      type: object
      properties:
        id:
          type: integer
        book:
          type: string
        chapter:
          type: integer
        verseNumber:
          type: integer
        text:
          type: string
        version:
          type: string
      required:
        - id
        - book
        - chapter
        - verseNumber
        - text
        - version
    DailyVerse:
      type: object
      properties:
        verse:
          $ref: "#/components/schemas/Verse"
        reflection:
          type: string
        date:
          type: string
      required:
        - verse
        - reflection
        - date
    VerseList:
      type: object
      properties:
        verses:
          type: array
          items:
            $ref: "#/components/schemas/Verse"
        total:
          type: integer
      required:
        - verses
        - total
    Devotional:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        verseReference:
          type: string
        verseText:
          type: string
        category:
          type: string
        readTime:
          type: integer
        date:
          type: string
      required:
        - id
        - title
        - content
        - verseReference
        - verseText
        - category
        - readTime
        - date
    DevotionalList:
      type: object
      properties:
        devotionals:
          type: array
          items:
            $ref: "#/components/schemas/Devotional"
        total:
          type: integer
      required:
        - devotionals
        - total
    Book:
      type: object
      properties:
        name:
          type: string
        testament:
          type: string
        chapters:
          type: integer
      required:
        - name
        - testament
        - chapters
    BookList:
      type: object
      properties:
        books:
          type: array
          items:
            $ref: "#/components/schemas/Book"
      required:
        - books