# lib/api-zod/src/generated/types/book.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bible Verse Daily API
 * OpenAPI spec version: 0.1.0
 */

export interface Book {
  name: string;
  testament: string;
  chapters: number;
}