# lib/api-zod/src/generated/types/bookList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bible Verse Daily API
 * OpenAPI spec version: 0.1.0
 */
import type { Book } from "./book";

export interface BookList {
  books: Book[];
}