# lib/api-zod/src/generated/types/dailyVerse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bible Verse Daily API
 * OpenAPI spec version: 0.1.0
 */
import type { Verse } from "./verse";

export interface DailyVerse {
  verse: Verse;
  reflection: string;
  date: string;
}