# lib/api-zod/src/generated/types/devotional.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bible Verse Daily API
 * OpenAPI spec version: 0.1.0
 */

export interface Devotional {
  id: number;
  title: string;
  content: string;
  verseReference: string;
  verseText: string;
  category: string;
  readTime: number;
  date: string;
}