# lib/api-zod/src/generated/types/devotionalList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bible Verse Daily API
 * OpenAPI spec version: 0.1.0
 */
import type { Devotional } from "./devotional";

export interface DevotionalList {
  devotionals: Devotional[];
  total: number;
}