# lib/api-zod/src/generated/types/verse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bible Verse Daily API
 * OpenAPI spec version: 0.1.0
 */

export interface Verse {
  id: number;
  book: string;
  chapter: number;
  verseNumber: number;
  text: string;
  version: string;
}